+++ packages/dnd/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@udecode/plate-dnd",
-  "version": "19.5.0",
+  "version": "19.7.0",
   "description": "Drag &amp; drop feature for Plate",
   "license": "MIT",
   "homepage": "https://plate.udecode.io",
@@ -19,7 +19,7 @@
   ],
   "types": "dist/index.d.ts",
   "dependencies": {
-    "@udecode/plate-core": "19.5.0",
+    "@udecode/plate-core": "19.7.0",
     "raf": "^3.4.1"
   },
   "peerDependencies": {
