+++ packages/dnd/src/components/DragHandle.tsx
@@ -8,6 +8,7 @@ import {
 export type PlateDragHandleProps = {
   element: TElement;
 } & HTMLPropsAs<'button'>;
+
 export const usePlateDragHandleProps = (
   props: PlateDragHandleProps
 ): HTMLPropsAs<'button'> => {
@@ -16,7 +17,9 @@ export const usePlateDragHandleProps = (
     onMouseDown: (e: any) => e.stopPropagation(),
   };
 };
+
 export const DragHandle = createComponentAs<PlateDragHandleProps>((props) => {
   const htmlProps = usePlateDragHandleProps(props);
+
   return createElementAs('button', htmlProps);
 });
