+++ packages/dnd/src/components/DraggableBlockToolbarWrapper.ts
@@ -1,0 +1,9 @@
+import {
+  AsProps,
+  createComponentAs,
+  createElementAs,
+} from '@udecode/plate-core';
+
+export const DraggableBlockToolbarWrapper = createComponentAs<AsProps<'div'>>(
+  (props) => createElementAs('div', props)
+);
