+++ packages/dnd/src/components/DraggableDropline.ts
@@ -5,6 +5,7 @@ import {
 } from '@udecode/plate-core';
 
 export type DraggableDroplineProps = {} & HTMLPropsAs<'div'>;
+
 export const useDraggableDroplineProps = (
   props: DraggableDroplineProps
 ): HTMLPropsAs<'div'> => {
@@ -13,9 +14,11 @@ export const useDraggableDroplineProps = (
     ...props,
   };
 };
+
 export const DraggableDropline = createComponentAs<DraggableDroplineProps>(
   (props) => {
     const htmlProps = useDraggableDroplineProps(props);
+
     return createElementAs('div', htmlProps);
   }
 );
