+++ packages/dnd/src/components/DraggableGutterLeftProps.ts
@@ -6,6 +6,7 @@ import {
 } from '@udecode/plate-core';
 
 export type DraggableGutterLeftProps = {} & HTMLPropsAs<'div'>;
+
 export const useDraggableGutterLeft = (
   props: DraggableGutterLeftProps
 ): HTMLPropsAs<'div'> => {
@@ -14,9 +15,11 @@ export const useDraggableGutterLeft = (
     ...props,
   };
 };
+
 export const DraggableGutterLeft = createComponentAs<AsProps<'div'>>(
   (props) => {
     const htmlProps = useDraggableGutterLeft(props);
+
     return createElementAs('div', htmlProps);
   }
 );
