+++ packages/dnd/src/components/useDraggableState.ts
@@ -1,7 +1,15 @@
-import { useRef } from 'react';
+import React, { useRef } from 'react';
+import { ConnectDragSource } from 'react-dnd';
 import { TElement } from '@udecode/plate-core';
 import { useDndBlock } from '../hooks';
-import { DraggableState } from './DraggableRoot';
+import { DropLineDirection } from '../types';
+
+export type DraggableState = {
+  dropLine: DropLineDirection;
+  isDragging: boolean;
+  rootRef: React.RefObject<HTMLDivElement>;
+  dragRef: ConnectDragSource;
+};
 
 export const useDraggableState = (props: {
   element: TElement;
@@ -15,4 +23,3 @@ export const useDraggableState = (props: {
 
   return { dropLine, isDragging, rootRef, dragRef };
 };
-``;
