--- packages/ui/dnd/src/components/index.ts
@@ -1,9 +1,0 @@
-/**
- * @file Automatically generated by barrelsby.
- */
-
-export * from './Draggable.styles';
-export * from './Draggable.types';
-export * from './PlateDraggable';
-export * from './grabberTooltipProps';
-export * from './withPlateDraggable';
