+++ packages/ui/dnd/src/index.ts
@@ -2,4 +2,8 @@
  * @file Automatically generated by barrelsby.
  */
 
-export * from './components/index';
+export * from './Draggable.styles';
+export * from './Draggable.types';
+export * from './PlateDraggable';
+export * from './grabberTooltipProps';
+export * from './withPlateDraggable';
